screensize & sprites

+++ drawsprite.c
@@ -39,8 +39,8 @@ void	draw_sprite_from_end(t_player *player, int i, int j, int count)
 	if (j < wall_h && i < wall_h &&
 	i >= 0 && j >= 0)
 	{
-	color = set_texture_sprite(player, (int)(j * player->ids.xpm_sprite_h / wall_h),
-	(int)(player->ids.xpm_sprite_w - (i * player->ids.xpm_sprite_w / wall_h)));
+	color = set_texture_sprite(player, j * player->ids.xpm_sprite_h / wall_h,
+	player->ids.xpm_sprite_w - (i * player->ids.xpm_sprite_w / wall_h));
 	if ((color > 0 && xposition_end - i >= 0 && xposition_end - i
 	< player->struct_screen.x && player->struct_screen.y / 2 - wall_h / 2 + j
 	>= 0 && player->struct_screen.y / 2 - wall_h / 2 + j <
@@ -61,8 +61,8 @@ void	draw_sprite_from_start(t_player *player, int i, int j, int count)
 	if (j < wall_h && i < wall_h &&
 	i >= 0 && j >= 0)
 	{
-	color = set_texture_sprite(player, (int)(j * player->ids.xpm_sprite_h / wall_h),
-	(int)(i * player->ids.xpm_sprite_w / wall_h));
+	color = set_texture_sprite(player, j * player->ids.xpm_sprite_h / wall_h,
+	i * player->ids.xpm_sprite_w / wall_h);
 	if ((color > 0 && xposition_start + i >= 0 && xposition_start + i
 	< player->struct_screen.x && player->struct_screen.y / 2 - wall_h / 2 + j
 	>= 0 && player->struct_screen.y / 2 - wall_h / 2 + j <
